feat: Add Breadcrumbs data BreadcrumbsPreprocessor
Adds a python script which processes the location data from the database
and the POIs and creates a timeseries of which POI a user is visiting by
associating it with the nearest POI to the user's location.

Adds utility functions for saving the data as a CSV and printing the
resulting graphs that can be generated.